cambios

+++ app/Http/Controllers/Usuarios/Usuarios.php
@@ -20,9 +20,17 @@ class Usuarios extends Controller
     {
         $this->middleware(['role_or_permission:Administrador|Usuarios']);
     }
-    public function index(UsuariosDataTable $dataTable)
+
+    public function index(UsuariosDataTable $dataTable,$rol=null)
     {
-        return $dataTable->render('usuarios.usuarios.index');
+        $roles=Role::all();
+        $data = array('roles' =>$roles );
+        $rol_m=Role::find($rol);
+        if($rol_m){
+            return $dataTable->with('rol',$rol)->render('usuarios.usuarios.index',$data);
+        }
+        return $dataTable->render('usuarios.usuarios.index',$data);
+        
     }
 
     public function nuevo()
@@ -106,6 +114,7 @@ class Usuarios extends Controller
     public function editarUsuario($idUsuario)
     {
         $usuario=User::findOrFail($idUsuario);
+        $this->authorize('editar', $usuario);
         $roles=Role::all();
         $provincias=Provincia::all();
         $data = array('usuario'=>$usuario,'roles' => $roles,'provincias'=>$provincias);
@@ -114,12 +123,58 @@ class Usuarios extends Controller
 
     public function actualizar(RqEditar $request)
     {
-        return $request;
+        $user=User::findOrFail($request->usuario);
+        $this->authorize('editar', $user);
+        $user->email=$request->email;
+        $user->name=$request->email;
+        
+        if($request->password){
+            $user->password=Hash::make($request->password);
+        }
+
+        $user->nombres=$request->nombres;
+        $user->apellidos=$request->apellidos;
+        $user->sexo=$request->sexo;
+        $user->tipo_identificacion=$request->tipo_identificacion;
+        $user->identificacion=$request->identificacion;
+        $user->fecha_nacimiento=$request->fecha_nacimiento;
+        $user->estado_civil=$request->estado_civil;
+        $user->etnia=$request->etnia;
+        $user->telefono=$request->telefono;
+        $user->celular=$request->celular;
+        $user->pais=$request->pais;
+        $user->parroquia_id=$request->parroquia;
+        $user->direccion=$request->direccion;
+        $user->tiene_discapacidad=$request->tiene_discapacidad;
+        $user->porcentaje_discapacidad=$request->porcentaje_discapacidad??0;
+        $user->tiene_carnet_conadis=$request->tiene_carnet_conadis;
+        $user->porcentaje_carnet_conadis=$request->porcentaje_carnet_conadis??0;
+        
+        $user->save();
+
+        if ($request->hasFile('foto')) {
+            if ($request->file('foto')->isValid()) {
+                Storage::delete($user->foto);
+                $extension = $request->foto->extension();
+                $path = Storage::putFileAs(
+                    'public/usuarios', $request->file('foto'), $user->id.'.'.$extension
+                );
+                $user->foto=$path;
+                $user->save();
+            }
+        }
+        $user->syncRoles($request->roles);
+        $request->session()->flash('success','Usuario actualizado');
+        
+        return redirect()->route('usuarios');
+
     }
 
     public function eliminar(Request $request, $idUsuario)
     {
+
         $user=User::findOrFail($idUsuario);
+        $this->authorize('eliminar', $user);
         $user->delete();
         $request->session()->flash('success','Usuario eliminado');
         return redirect()->route('usuarios');
